requirements.txt
grpcio
grpcio-tools
sqlite3